[data, model, recipe, training] feat: EuroVL sqrt per-token loss, packing finalize, vocab/rope fixes
Bundles the EuroVL working-branch changes:

- data: square-root per-token loss (InternVL3.5 eq.2) — EuroVLTaskEncoder bakes
  1/sqrt(N) into loss_mask (opt-in sqrt_loss_weighting flag); losses.py keeps
  num_tokens as float (no int truncation of the fractional weight sum).
- recipe: enable sqrt loss + calculate_per_token_loss=True; native CE fusion;
  rotary_base=1000000 and should_pad_vocab=True; energon packing wiring.
- data: energon fill-to-seq_length packing hooks + packing_buffer_size plumbing;
  drop epochs/compute_train_iters auto-derivation.
- training: setup honors should_pad_vocab for explicit vocab (TP>1 fix);
  vlm_step routes energon-packed batches through THD cu_seqlens.
- model: configuration_euro_vl rope_theta=1000000 / max_position 32768.

+++ src/megatron/bridge/training/losses.py
@@ -99,7 +99,11 @@ def masked_next_token_loss(
             fatal=False,
         )
 
-    num_tokens = loss_mask.sum().clone().detach().to(torch.int)
+    # Keep float (do not truncate to int): with fractional per-token loss weights
+    # (e.g. EuroVL sqrt loss) this is the sum of weights used as the global loss
+    # denominator, and truncating would bias normalization. For binary 0/1 masks the
+    # value is integer-valued, so this is identical to the previous behavior bar dtype.
+    num_tokens = loss_mask.sum().clone().detach()
     reporting_loss = torch.cat([loss.clone().detach().view(1), num_tokens.view(1)])
 
     return (loss, num_tokens, {"lm loss": reporting_loss})
